add feature order and orderDetail

+++ API_JAVASpringBoot/API_JAVA/src/main/java/api/java/entities/Shipper.java
@@ -80,7 +80,7 @@ public class Shipper implements Serializable {
 	}
 
 	@Column(name = "UserId")
-	public float getUserId() {
+	public int getUserId() {
 		return this.UserId;
 	}
